--- src/main/java/konkuk/chacall/domain/foodtruck/domain/Menu.java
@@ -1,33 +1,0 @@
-package konkuk.chacall.domain.foodtruck.domain;
-
-import jakarta.persistence.*;
-import konkuk.chacall.global.common.domain.BaseEntity;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "menus")
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-public class Menu extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "menu_id", nullable = false)
-    private Long menuId;
-
-    @Column(nullable = false)
-    private String name;
-
-    @Column(nullable = false)
-    private Integer price;
-
-    private String description;
-
-    private String imageUrl;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "food_truck_id", nullable = false)
-    private FoodTruck foodTruck;
-}
