+++ src/main/java/konkuk/chacall/domain/foodtruck/domain/model/AvailableDate.java
@@ -1,28 +1,32 @@
-package konkuk.chacall.domain.foodtruck.domain;
+package konkuk.chacall.domain.foodtruck.domain.model;
 
 import jakarta.persistence.*;
-import konkuk.chacall.domain.region.domain.model.Region;
 import konkuk.chacall.global.common.domain.BaseEntity;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.time.LocalDate;
+
 @Entity
-@Table(name = "food_truck_service_areas")
+@Table(name = "available_dates")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
-public class FoodTruckServiceArea extends BaseEntity {
+public class AvailableDate extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "food_truck_service_area_id", nullable = false)
-    private Long foodTruckServiceAreaId;
+    @Column(name = "available_date_id", nullable = false)
+    private Long availableDateId;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "region_id", nullable = false)
-    private Region region;
+    @Column(nullable = false)
+    private LocalDate startAt;
+
+    @Column(nullable = false)
+    private LocalDate endAt;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "food_truck_id", nullable = false)
     private FoodTruck foodTruck;
+
 }
