+++ src/main/java/konkuk/chacall/domain/foodtruck/domain/model/Menu.java
@@ -1,4 +1,4 @@
-package konkuk.chacall.domain.foodtruck.domain;
+package konkuk.chacall.domain.foodtruck.domain.model;
 
 import jakarta.persistence.*;
 import konkuk.chacall.global.common.domain.BaseEntity;
@@ -6,27 +6,28 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.time.LocalDate;
-
 @Entity
-@Table(name = "available_dates")
+@Table(name = "menus")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
-public class AvailableDate extends BaseEntity {
+public class Menu extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "available_date_id", nullable = false)
-    private Long availableDateId;
+    @Column(name = "menu_id", nullable = false)
+    private Long menuId;
 
     @Column(nullable = false)
-    private LocalDate startAt;
+    private String name;
 
     @Column(nullable = false)
-    private LocalDate endAt;
+    private Integer price;
+
+    private String description;
+
+    private String imageUrl;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "food_truck_id", nullable = false)
     private FoodTruck foodTruck;
-
 }
