+++ src/main/java/konkuk/chacall/domain/foodtruck/domain/repository/FoodTruckRepository.java
@@ -1,6 +1,6 @@
 package konkuk.chacall.domain.foodtruck.domain.repository;
 
-import konkuk.chacall.domain.foodtruck.domain.FoodTruck;
+import konkuk.chacall.domain.foodtruck.domain.model.FoodTruck;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.data.jpa.repository.JpaRepository;
