+++ src/main/java/konkuk/chacall/domain/member/domain/SavedFoodTruck.java
@@ -1,7 +1,7 @@
 package konkuk.chacall.domain.member.domain;
 
 import jakarta.persistence.*;
-import konkuk.chacall.domain.foodtruck.domain.FoodTruck;
+import konkuk.chacall.domain.foodtruck.domain.model.FoodTruck;
 import konkuk.chacall.domain.user.domain.model.User;
 import konkuk.chacall.global.common.domain.BaseEntity;
 import lombok.*;
