+++ src/main/java/konkuk/chacall/domain/member/presentation/dto/response/ReservationForRatingResponse.java
@@ -1,7 +1,7 @@
 package konkuk.chacall.domain.member.presentation.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
-import konkuk.chacall.domain.foodtruck.domain.FoodTruck;
+import konkuk.chacall.domain.foodtruck.domain.model.FoodTruck;
 import konkuk.chacall.domain.reservation.domain.model.Reservation;
 
 import java.util.List;
