+++ src/main/java/konkuk/chacall/domain/member/presentation/dto/response/SavedFoodTruckResponse.java
@@ -1,7 +1,7 @@
 package konkuk.chacall.domain.member.presentation.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
-import konkuk.chacall.domain.foodtruck.domain.FoodTruck;
+import konkuk.chacall.domain.foodtruck.domain.model.FoodTruck;
 
 public record SavedFoodTruckResponse(
         @Schema(description = "푸드트럭 식별자", example = "1")
